Validation Errors for Forms

+++ app/assets/javascripts/controllers/profile_controller.js.coffee
@@ -1,47 +1,45 @@
-ETahi.ProfileController = Ember.ObjectController.extend ETahi.FileUploadMixin,
-  hideAffiliationForm: true
-
-  errorText: ""
+ETahi.ProfileController = Ember.Controller.extend ETahi.FileUploadMixin, ETahi.ValidationErrorsMixin,
+  showAffiliationForm: false
 
   affiliations: Ember.computed.alias "model.affiliationsByDate"
 
   avatarUploadUrl: ( ->
-    "/users/#{@get('id')}/update_avatar"
+    "/users/#{@get('model.id')}/update_avatar"
   ).property('id')
 
   selectableInstitutions: (->
-    @get('institutions').map (institution) ->
+    @get('model.institutions').map (institution) ->
       id: institution
       text: institution
-  ).property('institutions')
+  ).property('model.institutions')
 
   actions:
     uploadFinished: (data, filename) ->
       @uploadFinished(data, filename)
-      @set('model.avatarUrl', data.avatar_url)
+      @set 'model.avatarUrl', data.avatar_url
 
-    toggleAffiliationForm: ->
-      @set('newAffiliation', @store.createRecord('affiliation'))
-      @toggleProperty('hideAffiliationForm')
-      false
+    hideNewAffiliationForm: ->
+      @clearValidationErrors()
+      @set 'showAffiliationForm', false
+      @get('newAffiliation').deleteRecord() if @get('newAffiliation.isNew')
+
+    showNewAffiliationForm: ->
+      @set 'newAffiliation', @store.createRecord('affiliation')
+      @set 'showAffiliationForm', true
 
     removeAffiliation: (affiliation) ->
-      if confirm("Are you sure you want to destroy this affiliation?")
-        affiliation.destroyRecord()
+      affiliation.destroyRecord() if confirm('Are you sure you want to destroy this affiliation?')
 
     commitAffiliation:(affiliation) ->
-      affiliation.set('user', @get('model'))
+      affiliation.set 'user', @get('model')
+      @clearValidationErrors()
+
       affiliation.save().then(
         (affiliation) =>
           affiliation.get('user.affiliations').addObject(affiliation)
-          @send('toggleAffiliationForm')
+          @send 'hideNewAffiliationForm'
         ,
-        (errorResponse) =>
-          affiliation.set('user', null)
-          errors = for key, value of errorResponse.errors
-            messages = for msg in value
-              # TODO: Use this in the other error handlers.
-              "#{key.dasherize().capitalize().replace("-", " ")} #{value}"
-            messages.join(", ")
-          Tahi.utils.togglePropertyAfterDelay(@, 'errorText', errors.join(', '), '', 5000)
+        (response) =>
+          affiliation.set 'user', null
+          @displayValidationErrorsFromResponse response
       )
